move functions to common module
