+++ test_sets/Defects/Defects_release_8_1.txt
@@ -1,6 +1,6 @@
 ./test_scripts/Defects/8_1/2989_RC_RPCs_and_disabled_policies.lua
 ./test_scripts/Defects/8_1/3800_Send_GetSystemCapability_during_OnSystemCapabilityUpdated.lua
 ./test_scripts/Defects/8_1/3805_crash_during_langchange.lua
-./test_scripts/Defects/8_1/3838_1_SDL_does_not_send_UI.ClosePopUp_in_case_VR.PerformInteraction_TIME_OUT.lua
-./test_scripts/Defects/8_1/3838_2_SDL_does_not_send_UI.ClosePopUp_in_case_VR.PerformInteraction_SUCCESS.lua
-./test_scripts/Defects/8_1/3838_3_SDL_does_not_send_UI.ClosePopUp_in_case_two_PerformInteraction_requests.lua
+./test_scripts/Defects/8_1/3838/3838_1_SDL_does_not_send_UI.ClosePopUp_in_case_VR.PerformInteraction_TIME_OUT.lua
+./test_scripts/Defects/8_1/3838/3838_2_SDL_does_not_send_UI.ClosePopUp_in_case_VR.PerformInteraction_SUCCESS.lua
+./test_scripts/Defects/8_1/3838/3838_3_SDL_does_not_send_UI.ClosePopUp_in_case_two_PerformInteraction_requests.lua
